게임 기본 로직 구현_4
- 공과 플레이어에 속도와 가속도 개념 적용

+++ NetworkGamePractice/src/main.cpp
@@ -34,7 +34,7 @@ int main()
 
 		while (tickTimer >= FRAME_TIME)
 		{
-			Game::Get().tick(tickTimer);
+			Game::Get().tick(static_cast<float>(tickTimer));
 			tickTimer -= FRAME_TIME;
 		}
